+++ libbookmarksmodule/tests/resources/Bookmarks
@@ -94,7 +94,7 @@
                      "date_modified": "13231269683027166",
                      "guid": "ca6bab0c-593c-4c72-8052-70784b2cd08e",
                      "id": "18",
-                     "name": "Subfolder2",
+                     "name": "Subfolder1",
                      "type": "folder"
                   },
                   {
